Add files via upload
AES (Advanced Encryption Standard) is a symmetric encryption algorithm used to securely encrypt and protect data. This algorithm is faster and more efficient than public-key (asymmetric) encryption methods. AES is a standard set by the United States National Institute of Standards and Technology (NIST).

AES is one of the block cipher algorithms, meaning it works in blocks of a certain size. It can work using 128, 192 or 256 bit long keys. As the key length increases, encryption security increases, but so does processing power and speed.

AES works using a series of mathematical operations. During the encryption process, operations such as XOR (perfect overlap) and transformations are performed between the plaintext (the text to be encrypted) and the keys. These operations are used to convert the plaintext into ciphertext. The decryption process performs the same operations in reverse to convert the ciphertext into plaintext.